+++ embedded-graphics/src/prelude.rs
@@ -1,14 +1,12 @@
 //! Prelude
 
+#[doc(no_inline)]
 pub use crate::{
-    drawable::{Drawable, Pixel},
     fonts::Font,
     geometry::{Dimensions, Point, Size},
     image::{ImageDimensions, IntoPixelIter},
     pixelcolor::{raw::RawData, GrayColor, IntoStorage, PixelColor, RgbColor},
     primitives::Primitive,
     transform::Transform,
+    DrawTarget, Drawable, Pixel,
 };
-
-#[doc(no_inline)]
-pub use crate::draw_target::DrawTarget;