+++ simulator/src/output_settings.rs
@@ -1,9 +1,5 @@
 use crate::{display::SimulatorDisplay, theme::BinaryColorTheme};
-use embedded_graphics::{
-    geometry::{Point, Size},
-    pixelcolor::PixelColor,
-    DrawTarget,
-};
+use embedded_graphics::prelude::*;
 
 /// Output settings.
 #[derive(Debug, PartialEq, Eq, Clone)]